update: 更新文件 napi_common.cpp

+++ interfaces/kits/napi_common/src/napi_common.cpp
@@ -1182,8 +1182,8 @@ bool GetAccountTypeInDlpProperty(napi_env env, napi_value jsObject, DlpProperty&
         return false;
     }
 
-    if ((type > static_cast<uint32_t>(ENTERPRISE_ACCOUNT)) ||
-        (type < static_cast<uint32_t>(INVALID_ACCOUNT))) {
+    if ((static_cast<uint32_t>(type) > static_cast<uint32_t>(ENTERPRISE_ACCOUNT)) ||
+        (static_cast<uint32_t>(type) < static_cast<uint32_t>(INVALID_ACCOUNT))) {
         DLP_LOG_ERROR(LABEL, "js get wrong owner account type.");
         return false;
     }
@@ -1670,8 +1670,8 @@ bool GetVectorAuthUser(napi_env env, napi_value jsObject, std::vector<AuthUserIn
         userInfo.permExpiryTime = static_cast<uint64_t>(time);
         int64_t type;
         if (!GetInt64ValueByKey(env, obj, "authAccountType", type) ||
-            (type > static_cast<uint32_t>(ENTERPRISE_ACCOUNT) ||
-            type < static_cast<uint32_t>(INVALID_ACCOUNT))) {
+            (static_cast<uint32_t>(type) > static_cast<uint32_t>(ENTERPRISE_ACCOUNT) ||
+            static_cast<uint32_t>(type) < static_cast<uint32_t>(INVALID_ACCOUNT))) {
             DLP_LOG_ERROR(LABEL, "js get type fail");
             resultVec.clear();
             return false;
